Refactor worker support
