+++ src/java/io/bazel/rulesscala/worker/BUILD
@@ -1,0 +1,10 @@
+java_library(name = "worker",
+  srcs = [
+      "Processor.java",
+      "GenericWorker.java",
+      ],
+  deps = [
+    "//src/java/com/google/devtools/build/lib:worker",
+  ],
+  visibility = ["//visibility:public"],
+)
