+++ src/java/io/bazel/rulesscala/worker/Processor.java
@@ -1,0 +1,7 @@
+package io.bazel.rulesscala.worker;
+
+import java.util.List;
+
+public interface Processor {
+  void processRequest(List<String> args) throws Exception;
+}
